refactor(protocol): split protocol.py into a protocol package
Break the 659-line modelship/openai/protocol.py into a protocol/ package
with one module per endpoint family (base, error, usage, chat, embeddings,
audio, images, raw). The package __init__ re-exports the full public surface
with the same __all__, so every existing
`from modelship.openai.protocol import ...` keeps working unchanged.

Behavior-preserving except ChatCompletionRequest.stop, which switches from a
bare mutable default ([]) to Field(default_factory=list) — equivalent under
pydantic but keeps RUF012 happy now that OpenAIBaseModel's pydantic ancestry
lives in another module.
